+++ R/print.dynSurv.R
@@ -3,7 +3,7 @@
 print.dynSurv <- function(x, digits = max(4, getOption("digits") - 4), ...) {
 
   if (!inherits(x, "dynSurv")) {
-    stop("Use only with 'dynSurv' objects.")
+    stop("Use only with 'dynSurv' objects.\n")
   }
 
   print(round(x$pred, digits = digits))